Initial Setup For Oculus Quest
Setup VR/Quest Optimisation Settings in Unity's Editor.
Downloaded Oculus Integration.
This can now be Used as a very basic template, that is pre-optimized for Oculus Quest Development.

ps.
If using this as a template, the next steps would be to go to the Oculus Developer Website and create a
 developer account if you haven't already, then create an app to get an AppID. Plug in that AppID into the platform settings in Unity ( Go to Oculus > Platform > Edit Settings).